Update rk4.m
May I enter the dragon...
First of all, thanks for the awesome repo, mate.
I am new to coding and a student at the moment. So in k1, k2, k3, k4, there is an error, and most websites have made the same mistake.  Everyone has actually missed an 'h' term in the function as you can notice.
Regards
HadesDMen
DMen leaves the chat